requirements.txt
attrs==20.3.0
aws-cdk.assets==1.116.0
aws-cdk.aws-apigateway==1.116.0
aws-cdk.aws-applicationautoscaling==1.116.0
aws-cdk.aws-autoscaling-common==1.116.0
aws-cdk.aws-certificatemanager==1.116.0
aws-cdk.aws-cloudformation==1.116.0
aws-cdk.aws-cloudwatch==1.116.0
aws-cdk.aws-codeguruprofiler==1.116.0
aws-cdk.aws-codestarnotifications==1.116.0
aws-cdk.aws-cognito==1.116.0
aws-cdk.aws-dynamodb==1.116.0
aws-cdk.aws-ec2==1.116.0
aws-cdk.aws-ecr==1.116.0
aws-cdk.aws-ecr-assets==1.116.0
aws-cdk.aws-efs==1.116.0
aws-cdk.aws-elasticloadbalancingv2==1.116.0
aws-cdk.aws-events==1.116.0
aws-cdk.aws-iam==1.116.0
aws-cdk.aws-kinesis==1.116.0
aws-cdk.aws-kms==1.116.0
aws-cdk.aws-lambda==1.116.0
aws-cdk.aws-logs==1.116.0
aws-cdk.aws-route53==1.116.0
aws-cdk.aws-s3==1.116.0
aws-cdk.aws-s3-assets==1.116.0
aws-cdk.aws-signer==1.116.0
aws-cdk.aws-sns==1.116.0
aws-cdk.aws-sqs==1.116.0
aws-cdk.aws-ssm==1.116.0
aws-cdk.cloud-assembly-schema==1.116.0
aws-cdk.core==1.116.0
aws-cdk.custom-resources==1.116.0
aws-cdk.cx-api==1.116.0
aws-cdk.region-info==1.116.0
boto3==1.18.9
botocore==1.21.9
cattrs==1.6.0
constructs==3.3.107
jmespath==0.10.0
jsii==1.32.0
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6